ç'a l'air mieux

+++ includes/mlx_utils.h
@@ -6,7 +6,7 @@
 /*   By: faventur <faventur@student.42mulhouse.fr>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/02 12:06:01 by faventur          #+#    #+#             */
-/*   Updated: 2022/10/07 18:40:25 by faventur         ###   ########.fr       */
+/*   Updated: 2022/10/07 19:42:42 by faventur         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -171,7 +171,7 @@ void		draw_line(mlx_image_t *img, t_vector2 start, t_vector2 finish,
 				uint32_t color);
 void		draw_vertical_line(mlx_image_t *img, t_vector draw_start,
 				uint32_t draw_end, uint32_t color);
-void		draw_texture(t_program *data, int x, float wall_height,
+void		draw_texture(t_program *data, float x, float wall_height,
 				int texture_pos_x, t_texture texture);
 void		draw_line_beta(t_program *data, t_vector2 start, t_vector2 finish,
 				t_texture texture);
